Move JAVA_OPTS to JAVA_OPTS_APPEND to preserve run-java.sh functionality
The run-java.sh script tries to autoconfigure JAVA_OPTS. This includes default jvm options and also a debugging configuration if JAVA_DEBUG is set. The autoconfiguration is suppressed if JAVA_OPTS is set from outside the script. By moving the Quarkus default JAVA_OPTS to JAVA_OPTS_APPEND the autoconfig works and the Quarkus options are appended after.

Signed-off-by: Björn Großewinkelmann <github@ankertaugasse.de>